sam-deploy/pyproject.toml
[project]
name = "sam-deploy"
version = "0.1.0"
description = "Optimized AWS SAM deployment project with Lambda aliases and HTTP API stage variables"
authors = [
    {name = "Kshitiz",email = "dzgrotechnologies@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.28.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0",
    "inquirer>=3.1.0",
]

[tool.poetry]
packages = [{include = "sam_deploy", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
